switched from double to float

+++ Assets/Scripts/Configs/GroupConfig.cs
@@ -11,19 +11,19 @@ public class GroupConfig : ScriptableObject
     public RandomConfig RandomConfig {get; private set;}
 
     [field: SerializeField]
-    public double MaxSpeed {get; private set;} = 117.71d;
+    public float MaxSpeed {get; private set;} = 117.71f;
 
     [field: SerializeField]
-    public double MinSpeed {get; private set;} = 0.07d; // in GFlops
+    public float MinSpeed {get; private set;} = 0.07f; // in GFlops
 
     [field: SerializeField]
     public int ConnectionInterval {get; private set;} = 20;
 
 
     [field: SerializeField]
-    public double ServerLatency {get; private set;}
+    public float ServerLatency {get; private set;}
 
     [field: SerializeField]
-    public double ServerBandwidth {get; private set;}
+    public float ServerBandwidth {get; private set;}
 
 }
