+++ Assets/Scripts/Models/TimeTickSystem.cs
@@ -5,9 +5,9 @@ public class TimeTickSystem : MonoBehaviour
 {
     public static TimeTickSystem Instance { get; private set; }
     public static event Action<int> OnTick;
-    private const double TICK_DURATION = 0.001d; // 200 ms
+    private const float TICK_DURATION = 0.001f; // 200 ms
     public int CurTick {get; private set;} = 0;
-    private double _tickTimer = 0;
+    private float _tickTimer = 0;
 
     private void Awake()
     {
